-warning

+++ far/DlgBuilder.hpp
@@ -274,7 +274,7 @@ class DialogBuilderBase
 		void UpdateBorderSize()
 		{
 			T *Title = &DialogItems[0];
-			int MaxWidth = MaxTextWidth();
+			intptr_t MaxWidth = MaxTextWidth();
 			Title->X2 = Title->X1 + MaxWidth + 3;
 			Title->Y2 = DialogItems [DialogItemsCount-1].Y2 + 1;
